check token is null go to  logout function

+++ server/controllers/query/userRemoves.js
@@ -4,7 +4,7 @@ const connection = config.connection;
 module.exports.userRemoves = (req) => {
 
   return new Promise((resolve, reject) => {
-    connection.query("DELETE FROM user WHERE id = ? ",[req],(err,res) => {
+    connection.query("DELETE FROM user WHERE id = ? AND role <> 'admin'",[req],(err,res) => {
             (err)?reject(err):resolve(res);
     });
   });
